refactor(acp): own built-in command migration
